+++ elasticsearch/_sync/client/_base.py
@@ -245,8 +245,8 @@ class BaseClient:
         self._max_retries: Union[DefaultType, int] = DEFAULT
         self._retry_on_timeout: Union[DefaultType, bool] = DEFAULT
         self._retry_on_status: Union[DefaultType, Collection[int]] = DEFAULT
-        self._retry_backoff_base: Union[DefaultType, int] = DEFAULT
-        self._retry_backoff_cap: Union[DefaultType, int] = DEFAULT
+        self._retry_backoff_base: Union[DefaultType, float] = DEFAULT
+        self._retry_backoff_cap: Union[DefaultType, float] = DEFAULT
         self._is_serverless = False
         self._verified_elasticsearch = False
         self._otel = OpenTelemetry()
